Revert "concluding merge"
This reverts commit 507a8c969ab7b4bb1299a9d43fe96912f16f2071

+++ server/src/main/java/com/meetup/service/impl/ArticleServiceImpl.java
@@ -70,7 +70,14 @@ public class ArticleServiceImpl implements IArticleService {
     public void postArticle(final ArticleCreationDTO articleCreationDTO,
         final String userLogin) {
         User user = userDao.findUserByLogin(userLogin);
-        articleDao.insertNewArticle(articleCreationDTO, user.getId());
+        if (user == null) {
+            throw new UserNotFoundException();
+        }
+        if (RoleProcessor.isSpeaker(user)) {
+            articleDao.insertNewArticle(articleCreationDTO, user.getId());
+        } else {
+            throw new SpeakerOperationNotAllowedException();
+        }
     }
 
     /**
@@ -82,7 +89,23 @@ public class ArticleServiceImpl implements IArticleService {
     @Override
     public void removeArticle(final int articleID,
         final String userLogin) {
-        articleDao.removeArticle(articleID);
+        User user = userDao.findUserByLogin(userLogin);
+        if (user == null) {
+            throw new UserNotFoundException();
+        }
+        Article currentArticle = articleDao.findArticleByID(articleID);
+        if (currentArticle == null) {
+            throw new ArticleNotFoundException();
+        }
+        if (RoleProcessor.isSpeaker(user)) {
+            if (user.getId() == currentArticle.getAuthorID()) {
+                articleDao.removeArticle(articleID);
+            } else {
+                throw new SpeakerOperationNotAllowedException();
+            }
+        } else {
+            throw new SpeakerOperationNotAllowedException();
+        }
     }
 
     /**
@@ -159,7 +182,19 @@ public class ArticleServiceImpl implements IArticleService {
      */
     @Override
     public List<CommentaryDisplayDTO> getCommentaries(final int articleID) {
-        return articleDao.getArticleCommentaries(articleID);
+        List<Commentary> commentaries = articleDao
+            .getArticleCommentaries(articleID);
+        List<CommentaryDisplayDTO> displayableCommentaries = new ArrayList<>();
+
+        for (Commentary commentary : commentaries) {
+            User user = userDao.findUserById(commentary.getAuthorID());
+            displayableCommentaries.add
+                (CommentaryDTOConverter.convertToCommentaryDisplayDTO(
+                    commentary,
+                    user
+                ));
+        }
+        return displayableCommentaries;
     }
 
     /**
@@ -170,15 +205,14 @@ public class ArticleServiceImpl implements IArticleService {
      * @param commentary Commentary.
      */
     @Override
-    public CommentaryDisplayDTO postCommentary(final int articleID,
-        final String userLogin,
+    public CommentaryDisplayDTO postCommentary(final int articleID, final String userLogin,
         final Commentary commentary) {
         User user = userDao.findUserByLogin(userLogin);
         int userID = user.getId();
         Commentary commentCreated =
-            articleDao.addCommentary(articleID, userID, commentary);
+                articleDao.addCommentary(articleID, userID, commentary);
         return CommentaryDTOConverter
-            .convertToCommentaryDisplayDTO(commentCreated, user);
+                .convertToCommentaryDisplayDTO(commentCreated, user);
     }
 
     /**
